+++ src/components/details-rating.js
@@ -1,10 +1,22 @@
 import AbstractComponent from "./abstract-component";
 
 class DetailsRating extends AbstractComponent {
-  constructor(poster, title) {
+  constructor(poster, title, userRating) {
     super();
     this._poster = poster;
     this._title = title;
+    this._userRating = userRating;
+  }
+
+  _getRatingScoreTemplate() {
+    return [...Array(9)].map((_, i) => {
+      const value = i + 1;
+      return `
+      <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="${value}" 
+      id="rating-${value}"
+      ${parseInt(this._userRating, 10) === value ? `checked` : ``}>
+      <label class="film-details__user-rating-label" for="rating-${value}">${value}</label>`.trim();
+    }).join(` `);
   }
 
   getTemplate() {
@@ -25,32 +37,33 @@ class DetailsRating extends AbstractComponent {
             <p class="film-details__user-rating-feelings">How you feel it?</p>
 
             <div class="film-details__user-rating-score">
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="1" id="rating-1">
-              <label class="film-details__user-rating-label" for="rating-1">1</label>
+            ${this._getRatingScoreTemplate()}
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="1" id="rating-1">-->
+<!--              <label class="film-details__user-rating-label" for="rating-1">1</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="2" id="rating-2">
-              <label class="film-details__user-rating-label" for="rating-2">2</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="2" id="rating-2">-->
+<!--              <label class="film-details__user-rating-label" for="rating-2">2</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="3" id="rating-3">
-              <label class="film-details__user-rating-label" for="rating-3">3</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="3" id="rating-3">-->
+<!--              <label class="film-details__user-rating-label" for="rating-3">3</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="4" id="rating-4">
-              <label class="film-details__user-rating-label" for="rating-4">4</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="4" id="rating-4">-->
+<!--              <label class="film-details__user-rating-label" for="rating-4">4</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="5" id="rating-5">
-              <label class="film-details__user-rating-label" for="rating-5">5</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="5" id="rating-5">-->
+<!--              <label class="film-details__user-rating-label" for="rating-5">5</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="6" id="rating-6">
-              <label class="film-details__user-rating-label" for="rating-6">6</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="6" id="rating-6">-->
+<!--              <label class="film-details__user-rating-label" for="rating-6">6</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="7" id="rating-7">
-              <label class="film-details__user-rating-label" for="rating-7">7</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="7" id="rating-7">-->
+<!--              <label class="film-details__user-rating-label" for="rating-7">7</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="8" id="rating-8">
-              <label class="film-details__user-rating-label" for="rating-8">8</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="8" id="rating-8">-->
+<!--              <label class="film-details__user-rating-label" for="rating-8">8</label>-->
 
-              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="9" id="rating-9">
-              <label class="film-details__user-rating-label" for="rating-9">9</label>
+<!--              <input type="radio" name="score" class="film-details__user-rating-input visually-hidden" value="9" id="rating-9">-->
+<!--              <label class="film-details__user-rating-label" for="rating-9">9</label>-->
 
             </div>
           </section>
